References/requirements.txt
numpy==1.24.4
pillow==10.2.0
